+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Add middleware to extract tenant id from X-Tenant-ID header (for non-JWT endpoints)
+
 ### Changed
 
 ### Deprecated